pyproject.toml
[project]
name = "social-media-video-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diskcache>=5.6.3",
    "fastapi[standard]>=0.122.0",
    "python-dotenv>=1.2.1",
    "yt-dlp>=2025.11.12",
]
